python/requirements.txt
Flask==0.9
Jinja2==2.6
Werkzeug==0.8.3
pyserial==2.6
wsgiref==0.1.2
